Add support for disable-strict-forbidden-file-verification
Enabling this option will allow an event to trigger a build if the event
contains BOTH one or more wanted file paths AND one or more forbidden
file paths.  In other words, with this option, the build will not get
triggered if the change contains only forbidden files, otherwise it will
get triggered.

Support for this was added in Plugin Version 2.16.0.